Adjust behavior of MobileAccessibility.updateTextZoom and MobileAccessibility.getTextZoom
Adjust behavior of MobileAccessibility.updateTextZoom and
MobileAccessibility.getTextZoom so they both return the correct updated
zoom value to a callback method.

Adjust behavior of isTalkBackRunning and isVoiceOverRunning so that
they return true only for the appropriate platform.

+++ src/android/com/phonegap/plugin/mobileaccessibility/AbstractMobileAccessibilityHelper.java
@@ -1,20 +1,41 @@
+/**
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ *
+*/
+
 package com.phonegap.plugin.mobileaccessibility;
 
 import android.view.ViewParent;
 
 public abstract class AbstractMobileAccessibilityHelper {
-	protected MobileAccessibility mMobileAccessibility;
-	protected ViewParent mParent;
-	public abstract void initialize(MobileAccessibility mobileAccessibility);
-	public abstract boolean isClosedCaptioningEnabled();
-	public abstract boolean isScreenReaderRunning();
-	public abstract boolean isTouchExplorationEnabled();
-	public abstract void onAccessibilityStateChanged(boolean enabled);
-	public abstract void onCaptioningEnabledChanged(boolean enabled);
-	public abstract void onTouchExplorationStateChanged(boolean enabled);
-	public abstract void addStateChangeListeners();
-	public abstract void removeStateChangeListeners();
-	public abstract void announceForAccessibility(CharSequence text);
-	public abstract int getTextZoom();
-	public abstract void setTextZoom(int textZoom);
+    protected MobileAccessibility mMobileAccessibility;
+    protected ViewParent mParent;
+    public abstract void initialize(MobileAccessibility mobileAccessibility);
+    public abstract boolean isClosedCaptioningEnabled();
+    public abstract boolean isScreenReaderRunning();
+    public abstract boolean isTouchExplorationEnabled();
+    public abstract void onAccessibilityStateChanged(boolean enabled);
+    public abstract void onCaptioningEnabledChanged(boolean enabled);
+    public abstract void onTouchExplorationStateChanged(boolean enabled);
+    public abstract void addStateChangeListeners();
+    public abstract void removeStateChangeListeners();
+    public abstract void announceForAccessibility(CharSequence text);
+    public abstract double getTextZoom();
+    public abstract void setTextZoom(double textZoom);
 }
