+++ src/android/com/phonegap/plugin/mobileaccessibility/IceCreamSandwichMobileAccessibilityHelper.java
@@ -1,3 +1,24 @@
+/**
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ *
+*/
+
 package com.phonegap.plugin.mobileaccessibility;
 
 import android.accessibilityservice.AccessibilityServiceInfo;
@@ -7,46 +28,46 @@ import android.view.accessibility.AccessibilityManager.AccessibilityStateChangeL
 
 @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
 public class IceCreamSandwichMobileAccessibilityHelper extends
-		DonutMobileAccessibilityHelper {
-	protected AccessibilityStateChangeListener mAccessibilityStateChangeListener;
-	
-	@Override
-	public boolean isScreenReaderRunning() {		
-		return mAccessibilityManager.getEnabledAccessibilityServiceList(AccessibilityServiceInfo.FEEDBACK_SPOKEN).size() > 0;
-	}
-	
-	@Override
-	public void addStateChangeListeners() {
-		if (mAccessibilityStateChangeListener == null) {
-			mAccessibilityStateChangeListener = new InternalAccessibilityStateChangeListener();
-		}
-		mAccessibilityManager.addAccessibilityStateChangeListener(mAccessibilityStateChangeListener);
-	}
-	
-	@Override
-	public void removeStateChangeListeners() {
-		mAccessibilityManager.removeAccessibilityStateChangeListener(mAccessibilityStateChangeListener);
-		mAccessibilityStateChangeListener = null;
-	}
-	
-	@Override
-	public int getTextZoom() {
-		return mWebView.getSettings().getTextZoom();
-	}
-	
-	@Override
-	public void setTextZoom(int textZoom) {
-		final int zoom = textZoom;
-		//Log.i("MobileAccessibility", "setTextZoom(" + zoom + ")");
-		mWebView.getSettings().setTextZoom(zoom);
-	}
-	
-	protected class InternalAccessibilityStateChangeListener 
-		implements AccessibilityStateChangeListener {
-		
-		@Override
-		public void onAccessibilityStateChanged(boolean enabled) {
-			mMobileAccessibility.onAccessibilityStateChanged(enabled);
-		}
-	}
+        DonutMobileAccessibilityHelper {
+    protected AccessibilityStateChangeListener mAccessibilityStateChangeListener;
+
+    @Override
+    public boolean isScreenReaderRunning() {
+        return mAccessibilityManager.getEnabledAccessibilityServiceList(AccessibilityServiceInfo.FEEDBACK_SPOKEN).size() > 0;
+    }
+
+    @Override
+    public void addStateChangeListeners() {
+        if (mAccessibilityStateChangeListener == null) {
+            mAccessibilityStateChangeListener = new InternalAccessibilityStateChangeListener();
+        }
+        mAccessibilityManager.addAccessibilityStateChangeListener(mAccessibilityStateChangeListener);
+    }
+
+    @Override
+    public void removeStateChangeListeners() {
+        mAccessibilityManager.removeAccessibilityStateChangeListener(mAccessibilityStateChangeListener);
+        mAccessibilityStateChangeListener = null;
+    }
+
+    @Override
+    public double getTextZoom() {
+        return mWebView.getSettings().getTextZoom();
+    }
+
+    @Override
+    public void setTextZoom(double textZoom) {
+        final double zoom = textZoom;
+        //Log.i("MobileAccessibility", "setTextZoom(" + zoom + ")");
+        mWebView.getSettings().setTextZoom((int) zoom);
+    }
+
+    protected class InternalAccessibilityStateChangeListener
+        implements AccessibilityStateChangeListener {
+
+        @Override
+        public void onAccessibilityStateChanged(boolean enabled) {
+            mMobileAccessibility.onAccessibilityStateChanged(enabled);
+        }
+    }
 }
